src/main/java/algorithms/leetcode/RotateImage.java
package algorithms.leetcode;

/**
 * @link <a href="https://leetcode.com/problems/rotate-image/description/">https://leetcode.com/problems/rotate-image/description/</a>
 * @date 17.03.2023
 */
public class RotateImage {

    public void rotate(int[][] matrix) {

        if (matrix == null || matrix.length == 0) {
            return;
        }

        // flip diagonally
        for (int i = 0; i < matrix.length; i++) {
            for (int j = 0; j < matrix.length - i; j++) {
                int temp = matrix[i][j];
                matrix[i][j] = matrix[matrix.length - j - 1][matrix.length - i - 1];
                matrix[matrix.length - j - 1][matrix.length - i - 1] = temp;
            }
        }

        // flip horizontally
        for (int i = 0; i < matrix.length / 2; i++) {
            for (int j = 0; j < matrix.length; j++) {
                int temp = matrix[i][j];
                matrix[i][j] = matrix[matrix.length - i - 1][j];
                matrix[matrix.length - i - 1][j] = temp;
            }
        }
    }
}
